+++ mrz-3-edition/tester.m
@@ -34,7 +34,7 @@ polynomialDerivativeEvaluationInPoint([3, 4, 8, 2, 9], 7);  % 4818
 %% chapter 2
 % algorithmsepsEstimation
 epsEstimation();  % 2.2204e-16 (Matlab R2016b in Linux 64bit)
-% baseComputation
+baseComputation();  % 1 (Matlab R2016b in Linux 64bit)
 % parabolaSolver
 
 % exercises