+++ mrz-3-edition/tester.m
@@ -35,7 +35,7 @@ polynomialDerivativeEvaluationInPoint([3, 4, 8, 2, 9], 7);  % 4818
 % algorithmsepsEstimation
 epsEstimation();  % 2.2204e-16 (Matlab R2016b in Linux 64bit)
 baseComputation();  % 1 (Matlab R2016b in Linux 64bit)
-% parabolaSolver
+parabolaSolver(1, 4, 1);  % -3.7321, -0.2679
 
 % exercises
 % maxMinFinder